decouple shared and microcode - moved token hashing methods to token.c
