+++ microasm/src/shared/table.c
@@ -3,29 +3,6 @@
 #include "shared/table.h"
 #include "shared/memory.h"
 
-// FNV-1a
-uint32_t tokenHash(void* value) {
-    Token* token = value;
-    uint32_t hash = 2166126261u;
-
-    for(int i = 0; i < token->length; i++) {
-        hash ^= TOKEN_GET(*token)[i];
-        hash *= 16777619;
-    }
-
-    return hash;
-}
-
-bool tokenCmp(void* a, void* b) {
-    Token* tokA = a;
-    Token* tokB = b;
-
-    if(tokA->length != tokB->length) {
-        return false;
-    }
-    return strncmp(TOKEN_GET(*tokA), TOKEN_GET(*tokB), tokA->length) == 0;
-}
-
 uint32_t strHash(void* value) {
     char* str = value;
     uint32_t hash = 2166126261u;
